splitup strds_raster_management
